+++ Project/Project/Data/Mappers/UserModels.swift
@@ -9,6 +9,18 @@
 import Foundation
 import ObjectMapper
 
+enum UserFields: String {
+    case id = "id"
+    case active = "active"
+    case email = "email"
+    case password = "password"
+    case token = "token"
+    case createdAt = "createdAt"
+    case updatedAt = "updatedAt"
+    case name = "name"
+    case role = "role"
+}
+
 class UserModel: Mappable {
     var id: Int?
     var active: Bool?
