--- Project/Project/Model/UserFields.swift
@@ -1,21 +1,0 @@
-//
-//  UserFields.swift
-//  Project
-//
-//  Created by Taqtile on 09/01/18.
-//  Copyright © 2018 Taqtile. All rights reserved.
-//
-
-import Foundation
-
-enum UserFields: String {
-    case id = "id"
-    case active = "active"
-    case email = "email"
-    case password = "password"
-    case token = "token"
-    case createdAt = "createdAt"
-    case updatedAt = "updatedAt"
-    case name = "name"
-    case role = "role"
-}
