--- Project/Project/Services/TemplateAPIHandler.swift
@@ -1,16 +1,0 @@
-//
-//  TemplateAPIHandler.swift
-//  Project
-//
-//  Created by Taqtile on 10/01/18.
-//  Copyright © 2018 Taqtile. All rights reserved.
-//
-
-import Foundation
-
-class TemplateAPIHandler {
-    //MARK: Static properties - Endpoints
-    static let loginEndpoint = "https://tq-template-node.herokuapp.com/authenticate"
-    static let userListEndpoint = "https://tq-template-node.herokuapp.com/users?"
-    static let userEndpoint = "https://tq-template-node.herokuapp.com/user/"
-}
